[QA-2194] Don't use unnecessary object store for all the mocked requests
This pull request removes the previous data structure `mockedRequests` which was holding information about all the requests that were mocked and how they are made. Instead, we can simply pass around this information as basic objects from function to function - there is no need passing around references in our use case.

Thus, the method `logMockedRequests` has been removed as well as this required such a global data structure. I do not think that this is a big loss though since it seems like no one used this method yet anyway.

This change will reduce memory consumption as well. I have tested it out locally and have to say though that the memory consumption even with the previous pattern wasn't that high - when calling mockRequest 1000 times and making a network request with a reasonably sized JSON, the memory rose to 5.94MB. This is not terrible, but with this change we will reduce the memory consumption entirely because nodeJS will take care of garbage collection of unused objects for us.

I have also renamed the method `expectNetworkRequestToNotHaveBeenMade` to `expectRequestToNotHaveBeenMade`. Due to this and the removal of a function, this will require a major version bump.