[QA-2289] Remove Contract concept
This renames all instances of "contract" to "example", as per discussion with @thomaschaplin. This pull request is not yet the task to update the README but I have already removed instances in the readme relating to the previously removed methods as it would have been unsensible to rename "contract" to "example" in these cases if I would remove them later on anyway.