Release 4 (#48)
* [QA-2289] Improve test coverage and reduce complexity (#41)

* [QA-2289] Improve test coverage and reduce complexity

This is in preparation for release 4.

* [QA-2289] Add new inspect method (#42)

* added inspect and test cases for bodys

* Add new inspect method

* Add new inspect method

* [QA-2289] Remove methods expectRequestMadeMatching and expectRequestMadeMatchingSchema (#43)

* [QA-2289] Remove methods expectRequestMadeMatching and expectRequestMadeMatchingSchema

As discussed with @thomaschaplin, this removes the the methods expectRequestMadeMatching and expectRequestMadeMatchingSchema available on the MockedRequest object. This is because the `inspect()` method will essentially replace these behaviours - it should be up to the user about how they want to make assertions on the data.

* remove unused dependencies

* [QA-2289] Remove Contract concept (#44)

* [QA-2289] Remove Contract concept

This renames all instances of "contract" to "example", as per discussion with @thomaschaplin. This pull request is not yet the task to update the README but I have already removed instances in the readme relating to the previously removed methods as it would have been unsensible to rename "contract" to "example" in these cases if I would remove them later on anyway.

* [QA-2289] Remove another unused dependency (#45)

* [QA-2289] Update readme for new inspect method in release 4 (#46)

* add nvmrc file (#47)

Co-authored-by: Rick Schubert <rick@tray.io>