Remove mockRequestFromExample
This pull request removes the `mockRequestFromExample` function. This function does not fit anymore with the current style of the repository given that we have decided that mock-inspect is an enabler of contract testing instead of implementing contract testing on its own.

In my current blog article (@thomaschaplin already read the draft; not yet published though) I outline how `mockRequest` on its own can be used in conjunction with contracts to create contract testing completely without `mockRequestFromExample`.

This new versions should be published as a major change which is why I updated the package version as well.