Remove unnecessary expect dependency (#74)
* Remove unnecessary expect dependency

This dependeny prevented the module to work with mocha. Reason: When running tests with jest, the dependency "expect" is already in your `node_modules` and thus node can use it even though it wasn't installed via package.json. But with mocha, this dependency is not in the `node_modules` tree.

I could have added `expect` as a dependency to package.json or removed it entirely. I decided to remove it as it was only used in one place - a leftover from the old days.

* Use strict equality check

Co-authored-by: rickschubert <rick.schubert@youview.com>